+++ src/components/App/App.tsx
@@ -1,14 +1,9 @@
 import React from 'react';
 import { AppRootProps } from '@grafana/data';
-import { PluginPropsContext } from '../../utils/utils.plugin';
-import { Routes } from '../Routes';
+import { Routes } from 'components/Routes/Routes';
 
 export class App extends React.PureComponent<AppRootProps> {
   render() {
-    return (
-      <PluginPropsContext.Provider value={this.props}>
-        <Routes />
-      </PluginPropsContext.Provider>
-    );
+    return <Routes />;
   }
 }
