--- src/components/App/index.tsx
@@ -1,1 +1,0 @@
-export * from './App';
