--- src/components/testIds.ts
@@ -1,8 +1,0 @@
-export const testIds = {
-  appConfig: {
-    container: 'data-testid ac-container',
-    apiKey: 'data-testid ac-api-key',
-    apiUrl: 'data-testid ac-api-url',
-    submit: 'data-testid ac-submit-form',
-  },
-};
