--- src/constants.ts
@@ -1,7 +1,0 @@
-import pluginJson from './plugin.json';
-
-export const PLUGIN_BASE_URL = `/a/${pluginJson.id}`;
-
-export enum ROUTES {
-  Explore = 'explore',
-}
