+++ src/module.ts
@@ -1,10 +1,4 @@
 import { AppPlugin } from '@grafana/data';
-import { App } from './components/App';
-import { AppConfig } from './components/AppConfig';
+import { App } from 'components/App/App';
 
-export const plugin = new AppPlugin<{}>().setRootPage(App).addConfigPage({
-  title: 'Configuration',
-  icon: 'cog',
-  body: AppConfig,
-  id: 'configuration',
-});
+export const plugin = new AppPlugin<{}>().setRootPage(App);
