--- src/pages/Explore/index.tsx
@@ -1,2 +1,0 @@
-export * from './LogExplorationPage';
-export * from './LogExploration';
