+++ src/utils/scenes.ts
@@ -1,6 +1,6 @@
 import { SceneObject, sceneGraph } from '@grafana/scenes';
 import { LiveTailControl } from 'components/Explore/LiveTailControl';
-import { LogExploration } from 'pages/Explore';
+import { LogExploration } from 'pages/Explore/LogExploration';
 
 export function getLiveTailControl(model: SceneObject): LiveTailControl | undefined {
   return sceneGraph
