--- src/utils/utils.routing.ts
@@ -1,6 +1,0 @@
-import { PLUGIN_BASE_URL } from '../constants';
-
-// Prefixes the route with the base URL of the plugin
-export function prefixRoute(route: string): string {
-  return `${PLUGIN_BASE_URL}/${route}`;
-}
